move _get_pipeline_info_map to Study
